+++ packages/picasso/src/Indicator/story/Default.example.tsx
@@ -5,6 +5,24 @@ const Example = () => (
   <>
     <Container bottom='medium'>
       <Container inline right='small'>
+        <Indicator color='light-grey' />
+      </Container>
+      <Typography inline size='medium'>
+        Pending item
+      </Typography>
+    </Container>
+
+    <Container bottom='medium'>
+      <Container inline right='small'>
+        <Indicator color='green' />
+      </Container>
+      <Typography inline size='medium'>
+        Completed item
+      </Typography>
+    </Container>
+
+    <Container bottom='medium'>
+      <Container inline right='small'>
         <Indicator color='red' />
       </Container>
       <Typography inline size='medium'>